+++ interfaces/model/src/payment/allocation.rs
@@ -1,12 +1,13 @@
+use bigdecimal::BigDecimal;
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Allocation {
     pub allocation_id: String,
-    pub total_amount: i32,
-    pub spent_amount: i32,
-    pub remaining_amount: i32,
+    pub total_amount: BigDecimal,
+    pub spent_amount: BigDecimal,
+    pub remaining_amount: BigDecimal,
     #[serde(skip_serializing_if = "Option::is_none", default)]
     pub timeout: Option<String>,
     pub make_deposit: bool,
@@ -15,9 +16,9 @@ pub struct Allocation {
 impl Allocation {
     pub fn new(
         allocation_id: String,
-        total_amount: i32,
-        spent_amount: i32,
-        remaining_amount: i32,
+        total_amount: BigDecimal,
+        spent_amount: BigDecimal,
+        remaining_amount: BigDecimal,
         make_deposit: bool,
     ) -> Allocation {
         Allocation {
