+++ interfaces/model/src/payment/rejection.rs
@@ -1,10 +1,11 @@
+use bigdecimal::BigDecimal;
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Rejection {
     pub rejection_reason: crate::payment::RejectionReason,
-    pub total_amount_accepted: i32,
+    pub total_amount_accepted: BigDecimal,
     #[serde(skip_serializing_if = "Option::is_none", default)]
     pub message: Option<String>,
 }
@@ -12,7 +13,7 @@ pub struct Rejection {
 impl Rejection {
     pub fn new(
         rejection_reason: crate::payment::RejectionReason,
-        total_amount_accepted: i32,
+        total_amount_accepted: BigDecimal,
     ) -> Rejection {
         Rejection {
             rejection_reason,
